+++ packages/polaris-viz-core/src/hooks/useSparkBar.ts
@@ -65,6 +65,16 @@ export function useSparkBar({
   const filteredData = removeNullValues(data[0]);
   const [defaultData] = data;
 
+  const hasInfiniteData = filteredData.some((value) => !Number.isFinite(value));
+
+  if (hasInfiniteData) {
+    throw new Error('Data must be finite');
+  }
+
+  if (!Number.isFinite(targetValue)) {
+    throw new Error('Target value must be finite');
+  }
+
   const yScale = scaleLinear()
     .range(calculateRange(defaultData.data, height))
     .domain([