+++ packages/polaris-viz/src/components/ComboChart/hooks/useDualAxisTicks.ts
@@ -37,6 +37,13 @@ export function useDualAxisTicks({data, drawableHeight}: Props) {
   const primaryAxis = axes[sourceOfTruthIndex === 0 ? 0 : 1];
   const secondaryAxis = axes[sourceOfTruthIndex === 0 ? 1 : 0];
 
+  if (
+    !Number.isFinite(primaryAxis.max) ||
+    !Number.isFinite(secondaryAxis.max)
+  ) {
+    throw new Error('Data must be finite');
+  }
+
   const initialYScaleValues = getInitialYScaleValues({
     drawableHeight,
     primaryAxis,